Create Geofences Controller and Implement endpoint to fetch all Geofences data.

+++ Fleet ManagementWebApplication/Models/CircleGeofence.cs
@@ -5,7 +5,7 @@
         public int ID { get; set; }
         public int GeofenceID { get; set; }
         public int Radius { get; set; }
-        public int Latitude { get; set; }
-        public int Longitude { get; set; }
+        public float Latitude { get; set; }
+        public float Longitude { get; set; }
     }
 }
